API Reference

WaveOperators.Circle

WaveOperators.Grid

WaveOperators.HorizontalLine

WaveOperators.IntegralEquation

WaveOperators.Rectangle

WaveOperators.Rectangle

WaveOperators.ShapeSlice

WaveOperators.Slab

WaveOperators.VerticalLine

WaveOperators.add_current!

WaveOperators.add_current!

WaveOperators.add_mode!

WaveOperators.compute_modes

WaveOperators.generate_G

WaveOperators.getindices

WaveOperators.is_in

WaveOperators.pos

WaveOperators.set_contrast!

WaveOperators.solve

WaveOperators.Circle — Method

Circle(position::T, radius) where T <: Tuple

Creates a Circle at position with radius radius.

WaveOperators.Grid — Method

Grid(height::Real, width::Real, h::Real, k::Real)

Creates a height by width grid with discretization h and wavenumber k. This grid will define our problem domain.

WaveOperators.HorizontalLine — Type

HoriztonalLine(y2::Float64)

Creates a horiztonal line (across the entire domain) at y position y.

WaveOperators.IntegralEquation — Type

IntegralEquation

This object represents the physics equation

\[z + G\mathbf{diag}(\theta)z = Gb,\]

in compact form. Specifically, G_design maps from the design region to the design region, and G_target maps from the design region to the target. By only dealing with the reduced form of the system, we avoid unnecessary computation. The reduced form of the physics is

\[\begin{aligned} z^\mathrm{design} + G^\mathrm{design}z^\mathrm{design} &= b^\mathrm{design}\\ z^\mathrm{target} + G^\mathrm{target}z^\mathrm{design} &= b^\mathrm{target} \end{aligned}\]

WaveOperators.Rectangle — Method

Rectangle(w::N, h::M) where {N <: Real, M <: Real}

Creates a Rectangle at (0,0) with dimensions w by h.

WaveOperators.Rectangle — Method

Rectangle(position::T, shape::U) where {T <: Tuple, U <: Tuple}

Creates a Rectangle at positionwith dimensionsshape. Both inputs are tuples of floats (Tuple{Float64, Float64}`).

WaveOperators.ShapeSlice — Type

ShapeSlice(shape::Shape, line::VerticalLine)

Creates a vertical slice (defined by line) of shape.

WaveOperators.Slab — Type

Slab(position::Float64, width::Float64)

Creates a Slab (horizontal rectangle that strectes across the entire domain) at y position position with (vertical) width width.

WaveOperators.VerticalLine — Type

VerticalLine(x::Float64)

Creates a vertical line (across the entire domain) at x position x.

WaveOperators.add_current! — Method

add_current!(g::Grid, positions, input)

Adds excitation input to the grid g at positions positions.

WaveOperators.add_current! — Method

add_current!(g::Grid, s::S, input::N) where {S <: Shape, N <: Number}

Adds excitation input to the grid g at position s.

WaveOperators.add_mode! — Method

add_mode!(g::Grid, x, n_mode::Int)

Adds mode number n_mode to the grid g at position x, which may be an x axis position or a VerticalLine.

WaveOperators.compute_modes — Method

compute_modes(g::Grid, x)

Returns the modes of the excitation on the grid g.

WaveOperators.generate_G — Method

function generate_G(g::Grid, design, target)

Generates the IntegralEquation (Green's function G and excitations b) for the equation

\[z + G\mathbf{diag}(\theta)z = Gb,\]

The inputs design and target may be indices or shapes.

WaveOperators.getindices — Method

getindices(g::Grid, s::Shape) where S <: Shape

Gets indices (wrt the grid g) of the shape s

WaveOperators.is_in — Function

is_in(s::Shape, p::Position)

Checks if p (a tuple (x::Float64, y::Float64)) is in the shape s.

WaveOperators.pos — Function

pos(g::Grid, x)

Returns the position of x in grid

WaveOperators.set_contrast! — Method

set_contrast!(g::Grid, s::S, val) where S <: Shape

Sets the contrast of the shape s in grid g to be val.

WaveOperators.solve — Method

solve(g::Grid)

Computes the field on the grid g (an excitation should be added first).